Have InputPane update itself through an IContentProvider
Instead of making the controller of an inputpane responsible for
providing lines when they scroll into view, the InputPane will now use a
provided IContentProvider to request the lines of content it needs when
it needs them.
Diff3ContentProvider implements the IContentProvider interface using the
Diff3LineArray and a ILineProvider.