Redesigned navigation and editing in the presence of multi-column chars
Navigation is now column based and can only go to the first column of a
multi-column character instead of anywhere in the middle. Editing is
character based, moving content by multiple columns if a multi-column
character is deleted.
Also fixed tab characters to take up the space upto the next tabstop.