Split ContentEditor into SectionNavigator and ContentEditor
The SectionNavigator that navigates between merge result sections
needs to talk to HighlightAddingContentProviders for both input and
output panes.
The ContentEditor does not need this, but instead needs to know whether
a line from the ContentProvider is editable. Because this is specific to
the MergeResultContentProvider it made no sense adding such a method to
the IContentProvider interface. By splitting off the SectionNavigator we
can avoid it depending on MergeResultContentProvider.